move form into its own file, and organize files generally
